Use new address book upload endpoint
The new address book upload endpoint does not return suggestions of people you might know. However, we do not display people you might know anywhere in the client anymore, so we can safely set it to an empty set.
The code to handle people you might know is not being removed completely yet because we are still debating whether it will come back soon.