Port branch from old repository
This PR ports the branch `feature/add-event-decoder-ZIOS-5318` to the new repository.
This adds a separate database on the sync-engine which is used to store received update events before they have been decrypted.
This database ensures the decrypted events are not lost in case of a crash before processing them (as they can only be decrypted once).