Parse suggestions from address book search result
This restores the behaviour that was present at version: https://github.com/wireapp/wire-ios-sync-engine/blob/a84e152d82fde41e5a961eb2cf505c067150680c/Source/Synchronization/ZMAddressBookSync.m#L153 and then removed. Not that this is not the final implementation of the feature: that behaviour was not as designed, and will eventually be fixed (soonish).

With this pull request we have a similar issue that will be addressed later. The issue is that, since we search in batches, and take the response of a batch as the new list of suggestions, at any given time the suggestions will be the ones coming from the last batch upload. If the user has more contacts than the batch size (will upload more than one batch), the suggestions will change constantly after every batch upload. We could append the new results, instead of replacing them, but this will cause the list to grow potentially indefinitely, as we then never clean it.

It was decided to live with this limitation for now.