Cleaned up the notification creation (by @basine)
This is a squash of a series of commit by @basine.

(1) For events that create messages: Process the messages instead of the raw events to avoid double parsing of the event payload
(2) For call events: Move the responsibility of call state tracking (who initiated the call, what kind of call is it, did the selfUser join etc. into a separate object)
(3) For all events that don't create a message: Fetch and pass in the conversation before creating the event to avoid multiple fetches when copying events.