Mark the user profile update observer methods as optional. (#137)
The `UserProfileUpdateObserver` protocol contains a couple of methods relevant for updating emails, phone numbers and @names, however most of the conforming types do not handle all of these cases which leads to a lot of empty implementations just to conform to this protocol. This is why the methods of the protocol are now marked as `optional` (and `@objc` unfortunately as there are no optional conformances in Swift yet).