Add answered case to CallState
When you received an incoming call we didn't have state for when
the user had accepted the call. AVS will also add this so that
wcall_get_state returns the same state.

The userId has also been marked as optional since AVS does not
always send it in their callback handlers.