Fix for outgoing callkit calls
Outgoing callkit calls broke early during the calling 3 refactoring. When doing an outgoing
callkit you must request to start a call and only when asked to actually start the call.
This required some re-factoring of the interfaces to support this for both V2 and V3.

Most of the changes are actually from this re-factoring of the intefaces.
The VoiceChannelRouter is now an internal class (not really because of objc) and the UI
should only see the VoiceChannel protocol.