Fix call APNS after app is killed
When we receive a call after having killed the app we didn't display incoming
call notification because AVS wasn't ready to process calling messages. The
reason for this is that we weren't sending the calls/config request which AVS
needs to complete before it can do anything.

This PR allows this request to be made at all times. Previously it wasn't
made since the push channel wasn't open.