Add call snapshots
We are adding call snapshots so that we don't have to query AVS about the previous
call state. This means we will no longer call the following methods:

wcall_get_state
wcall_is_video_call
wcall_iterate_state

The reason for this change is that the above mentioned methods do not return correct
values at all times. In addition to that AVS is not happy about how frequently we
call them atm.