Ensure we delete teams we did not receive a deletion event for
It can happen that the user is offline for longer than 4 weeks, in this case said user will miss events and we will perform a slow sync.
In the case the user got removed from a team that the user already has locally in that period, we want to ensure that we delete that team from the client.
The `/teams` request will only return the teams the user is still in of course, so we fetch all local teams and mark them as `needsToBeUpdatedFromBackend`,
we reset this flag for all teams we fetch during the slow sync, (we just fetched them so there is no need to do it again after the slow sync completed).
After the slow sync we will try to fetch the deleted teams and receive a 4xx response and delete the team locally.