Logout user on authentication error / team deletion
Now that we support logout we can close the user session when the self client
gets deleted or the cookie expires. In order to help the user to sign in again
we supply the previous credentials in the authentication error user info.

When user gets removed from a team we also delete the account data after logging
out the user.