Fixes for account deletion & logout
Since logout is now asynchronous the account deleted needed to be refactored.

I also deleted `sessionManagerCreated(unauthenticatedSession:)` since didn't have much
of purpose anymore and was interfering with the logout. This could happen if it was
executed before the `sessionManagerDidFailToLogin()` or `sessionManagerWillLogout()`,
we would then lose the error message.

This should fix the pre-filling of the email on logout.