Workaround: OCMock corrupts memory when running sync strategy tests
ZMSyncStrategyTests uses a huge amount of mocking. To check for some sync states the ApplicationStatusDirectory is mocked together with some statuses. However after building with Xcode9 it somehow corrupts memory. The first variable of ApplicationStatusDirectory is set to pointer value 0x1 and any calls it to it crash the test suite. Changing the order of properties corrupts the apnsConfirmationStatus pointer value, but this is not used so test passes.