don’t require non active app state to handle call action
on iOS 11, tapping a call notification would call this method with an active app state, meaning we would not create and process the pending notifcation. Allowing an active state is acceptable because this method is only called when tapping a system notifcation, which can only be done when the app is not in the foreground.