remove uneccessary test
now we don’t require a non active app state when handling local notifications, this test was breaking because “showConversationListForUserSession” was being called. However, after some discussion with Jacob, we concluded that this test is outdated and no longer reflects a realistic situation because we wouldn’t have a user session if we were not already logged in.