configure callkit push string separately
The only time we want to include the conversation name in the push string is for the callkit overlay (since the conversation name would be in the title of the push notification/chathead). For all other push strings, we don’t add the conversation name to the localized string’s format arguments, so we need to handle callkit push strings manually.