Fix legal hold not being updated in conversations after restoring from backup
When a self client is discovered to be deleted during startup or login we were simply deleting it without any side effects and thus
not re-calculating the legal hold status in conversations. This can be fixed by deleting clients with the already exising
method deleteClientAndEndSession, which updates the legal hold status in the relevant conversations.